Notes and Coding Updates
Vincent helped me with the coding, and I used the notes to decipher what he did. Instead of creating new lines by inserting it blindly, he showed me to utilizes functions and to create things. Such as drawtable to rewrite the codes in index when clicking addActivity, creating the activity list, and anchoring checkboxes.